// frontend/src/interfaces/InputAcademicStaff.ts
export interface InputAcademicStaffInterface {

  user_id: number;
  academic_position: string;  
  first_name?: string;
  last_name?: string;
  age: number;
  birthday?: string;
  university_id?: number;
  faculty_id?: number;
  program_id?: number;
  address_id: number;
  gender_id?: number;
  contact_id:number;

}